Change paramaters to LocalMinima so that it will accept local minima within slightly less than the frame period
When confirming the trial start frames reported in the output matrix against the galvo + trial timer signals,  I found that one of the local minima in the galvo signal was not detected up by LocalMinima. Apparently this was because I had specified that LocalMinima should not include any local minimum within P samples of the previous local minimum, where P was the exact theoretical number of samples per frame. However, it turns out that there's some variability in the true number of samples per frame, so LocalMinimum was omitting minima that happened to be even slightly less that P samples after the previous local minimum. Thus, I changed P to P * 0.90, and this recovered the local minimum that had been previously omitted.